move icons logic into its own set of files
