+++ src/content/reducers/app.js
@@ -1,9 +1,8 @@
 // @flow
 import { combineReducers } from 'redux';
-import { createSelector } from 'reselect';
 
 import type { Action } from '../actions/types';
-import type { State, AppState, Reducer, IconWithClassName } from './types';
+import type { State, AppState, Reducer } from './types';
 
 function view(state: string = 'INITIALIZING', action: Action) {
   switch (action.type) {
@@ -27,35 +26,10 @@ function isURLSetupDone(state: boolean = false, action: Action) {
   }
 }
 
-function classNameFromUrl(url) {
-  return url.replace(/[/:.+>< ~()#,]/g, '_');
-}
 
-function favicons(state: Set<string>, action: Action) {
-  switch (action.type) {
-    case 'ICON_HAS_LOADED':
-      return new Set([...state, action.icon]);
-    case 'ICON_IN_ERROR': // nothing to do
-    default:
-      return state;
-  }
-}
-
-
-const appStateReducer: Reducer<AppState> = combineReducers({ view, isURLSetupDone, favicons });
+const appStateReducer: Reducer<AppState> = combineReducers({ view, isURLSetupDone });
 export default appStateReducer;
 
 export const getApp = (state: State): AppState => state.app;
 export const getView = (state: State): string => getApp(state).view;
 export const getIsURLSetupDone = (state: State): boolean => getApp(state).isURLSetupDone;
-export const getIcons = (state: State): Set<string> => getApp(state).favicons;
-export const getIconForNode = (state: State, node): string => getIcons(state).has(node.icon) ? node.icon : null;
-export const getIconClassNameForNode = createSelector(
-  getIcons, (state, node) => node,
-  (icons, node) => (icons.has(node.icon) ? classNameFromUrl(node.icon) : null)
-);
-export const getIconsWithClassNames: (State => IconWithClassName[]) = createSelector(
-  getIcons,
-  icons => [...icons].map(icon => ({ icon, className: classNameFromUrl(icon) }))
-);
-
